Enforce compilation warnings are errors on Windows

+++ test/comprehensive/host/testarray.cpp
@@ -89,10 +89,10 @@ void test_array_edl_ecalls(oe_enclave_t* enclave)
     {
         uint8_t data1[IV_SIZE];
         for (uint32_t i = 0; i < IV_SIZE; ++i)
-            data1[i] = i;
+            data1[i] = static_cast<uint8_t>(i);
         uint8_t data2[EXT_IV_SIZE];
         for (uint32_t i = 0; i < EXT_IV_SIZE; ++i)
-            data2[i] = i;
+            data2[i] = static_cast<uint8_t>(i);
 
         OE_TEST(ecall_named_dims(enclave, data1, data2) == OE_OK);
     }
